refactor: consolidate schema files into a single schemas.ts
Merge 6 individual schema files from src/state/schemas/ into a single
src/state/schemas.ts with one zod import. Move schemas.test.ts up one
level with updated import, update validator.ts import path, and delete
the now-empty schemas/ directory.
